Removed case-sensitivity from Master stencil naming of individual items.
If a file is called 'Master.png', it will now show up as 'Master' in the stencil instead of 'master'.
If a file has a space, like 'Hello World.png', it will now show up as 'Hellow World' in the stencil.

NOTE: That there is no special character removal added.